# Cache instantiated Python modules