I'm making this change to support some upcoming changes that will optimize the `pl-code` element to cache expensive-to-compute data. However, even on its own, this change already yields some nice improvements. I tested these changes on the `element/multipleChoice` question in the example course with five trials. I ran these tests with the `process-questions-in-worker` feature flag enabled.

Master:

```
freeform.generate: 0.74ms
freeform.prepare: 38.25ms
freeform.render: 35.43ms
freeform.generate: 0.55ms
freeform.prepare: 44.61ms
freeform.render: 41.02ms
freeform.generate: 0.27ms
freeform.prepare: 37.71ms
freeform.render: 38.84ms
freeform.generate: 0.34ms
freeform.prepare: 38.50ms
freeform.render: 32.91ms
freeform.generate: 0.49ms
freeform.prepare: 43.66ms
freeform.render: 36.21ms
```

On this branch:

```
freeform.generate: 0.77ms
freeform.prepare: 18.27ms
freeform.render: 19.76ms
freeform.generate: 0.65ms
freeform.prepare: 18.12ms
freeform.render: 23.98ms
freeform.generate: 0.11ms
freeform.prepare: 10.44ms
freeform.render: 16.50ms
freeform.generate: 0.59ms
freeform.prepare: 19.64ms
freeform.render: 22.99ms
freeform.generate: 0.62ms
freeform.prepare: 20.08ms
freeform.render: 25.06ms
```

Summary (blame ChatGPT for any wrong numbers):

| operation | `master` average | branch average | diff |
|--------|--------|--------|--------|
| `generate` | 0.48ms | 0.55ms | -14% |
| `prepare` | 40.55ms | 17.31ms | 57% |
| `render` | 36.88ms | 21.66ms | 41% |

`generate` is already very fast and doesn't use elements to begin with, so I'm inclined to say that the difference here is just due to normal variation.

Note that these changes will really only help with questions that use a given element multiple times. The improvement this will yield is proportional to the number of times an element is reused.